move helper function to xrptestutils
