+++ test/XRP/issued-currency-client-integration.test.ts
@@ -1,13 +1,10 @@
 import { assert } from 'chai'
-import { Wallet, WalletFactory, XrplNetwork, XrpUtils } from 'xpring-common-js'
-import TransactionStatus from '../../src/XRP/shared/transaction-status'
+import { WalletFactory, XrplNetwork } from 'xpring-common-js'
 import { XrpError } from '../../src/XRP'
 import IssuedCurrencyClient from '../../src/XRP/issued-currency-client'
-import GrpcNetworkClient from '../../src/XRP/network-clients/grpc-xrp-network-client'
 
 import XRPTestUtils from './helpers/xrp-test-utils'
-import { LedgerSpecifier } from '../../src/XRP/Generated/node/org/xrpl/rpc/v1/ledger_pb'
-import { AccountRootFlag, TransactionResult } from '../../src/XRP/shared'
+import { AccountRootFlag } from '../../src/XRP/shared'
 
 // A timeout for these tests.
 // eslint-disable-next-line @typescript-eslint/no-magic-numbers -- 1 minute in milliseconds
@@ -25,45 +22,6 @@ const issuedCurrencyClient = IssuedCurrencyClient.issuedCurrencyClientWithEndpoi
   rippledJsonUrl,
   XrplNetwork.Test,
 )
-
-async function runFlagIntegrationTestSuccess(
-  wallet: Wallet,
-  result: TransactionResult,
-  accountRootFlag: number,
-): Promise<void> {
-  // THEN the transaction was successfully submitted and the correct flag was set on the account.
-  const transactionHash = result.hash
-  const transactionStatus = result.status
-
-  // get the account data and check the flag bitmap
-  const networkClient = new GrpcNetworkClient(rippledGrpcUrl)
-  const account = networkClient.AccountAddress()
-  const classicAddress = XrpUtils.decodeXAddress(wallet.getAddress())
-  account.setAddress(classicAddress!.address)
-
-  const request = networkClient.GetAccountInfoRequest()
-  request.setAccount(account)
-
-  const ledger = new LedgerSpecifier()
-  ledger.setShortcut(LedgerSpecifier.Shortcut.SHORTCUT_VALIDATED)
-  request.setLedger(ledger)
-
-  const accountInfo = await networkClient.getAccountInfo(request)
-  if (!accountInfo) {
-    throw XrpError.malformedResponse
-  }
-
-  const accountData = accountInfo.getAccountData()
-  if (!accountData) {
-    throw XrpError.malformedResponse
-  }
-
-  const flags = accountData.getFlags()?.getValue()
-
-  assert.exists(transactionHash)
-  assert.equal(transactionStatus, TransactionStatus.Succeeded)
-  assert.isTrue(AccountRootFlag.checkFlag(accountRootFlag, flags!))
-}
 
 describe('IssuedCurrencyClient Integration Tests', function (): void {
   // Retry integration tests on failure.
@@ -133,8 +91,9 @@ describe('IssuedCurrencyClient Integration Tests', function (): void {
     )
 
     // THEN the transaction was successfully submitted and the correct flag was set on the account.
-    await runFlagIntegrationTestSuccess(
+    await XRPTestUtils.verifyFlagModification(
       wallet,
+      rippledGrpcUrl,
       result,
       AccountRootFlag.LSF_REQUIRE_AUTH,
     )
@@ -147,8 +106,9 @@ describe('IssuedCurrencyClient Integration Tests', function (): void {
     const result = await issuedCurrencyClient.disallowIncomingXrp(wallet)
 
     // THEN the transaction was successfully submitted and the correct flag was set on the account.
-    await runFlagIntegrationTestSuccess(
+    await XRPTestUtils.verifyFlagModification(
       wallet,
+      rippledGrpcUrl,
       result,
       AccountRootFlag.LSF_DISALLOW_XRP,
     )
